requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
email-validator>=2.1.0
dapr>=1.12.0
pytz>=2023.3.post1
python-dateutil>=2.8.2
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
openai>=1.12.0
tenacity>=8.2.3
cohere>=4.0.0
PyJWT>=2.8.0         # <-- add this if using import jwt
mangum>=0.17.0
